Modularize Lottie AST format models into dedicated domain packages
### Impact Summary
Restructures internal AST data models and serializers in the Lottie module from monolithic definition files into fine-grained, domain-specific packages (`format.values`, `format.properties`, `format.graphicelement`, `format.layer`). This change has zero external runtime impact and preserves exact behavioral parity across all animation loading, serialization, and rendering tests.

### Motivation
The monolithic format model files (`Properties.kt`, `Shapes.kt`, `Layers.kt`, `Values.kt`) grouped disparate concerns and polymorphic serializers together, making downstream maintenance, feature additions (such as advanced keyframing, gradients, and shape trimming), and isolated testing cumbersome. Breaking the models into modular files improves cohesion and prepares the codebase for incremental feature migrations.
