+++ remotecompose/lottie/src/main/java/com/google/android/horologist/remotecompose/lottie/format/graphicelement/grouping/Group.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2026 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.horologist.remotecompose.lottie.format.graphicelement.grouping
+
+import com.google.android.horologist.remotecompose.lottie.format.graphicelement.GraphicElement
+import com.google.android.horologist.remotecompose.lottie.format.graphicelement.ShapeType
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+/** A group of other graphic elements. This allows transforms to be nested. */
+@Serializable
+internal data class Group(
+  @SerialName("nm") override val name: String? = "",
+  @SerialName("hd") override val hidden: Boolean? = false,
+  @SerialName("ty") override val type: ShapeType = ShapeType.Group,
+  @SerialName("np") val numberOfProperties: Int? = null,
+  @SerialName("it") val shapes: List<GraphicElement>,
+) : GraphicElement
