+++ remotecompose/lottie/src/main/java/com/google/android/horologist/remotecompose/lottie/format/graphicelement/styles/Fill.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2026 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.horologist.remotecompose.lottie.format.graphicelement.styles
+
+import com.google.android.horologist.remotecompose.lottie.format.graphicelement.GraphicElement
+import com.google.android.horologist.remotecompose.lottie.format.graphicelement.ShapeType
+import com.google.android.horologist.remotecompose.lottie.format.properties.BaseScalarProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.StaticColorProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.StaticScalarProperty
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+/** Solid fill color */
+@Serializable
+internal data class Fill(
+  @SerialName("nm") override val name: String? = "",
+  @SerialName("hd") override val hidden: Boolean? = false,
+  @SerialName("ty") override val type: ShapeType = ShapeType.Fill,
+  @SerialName("o") val opacity: BaseScalarProperty = StaticScalarProperty(value = 100f),
+  @SerialName("c") val color: StaticColorProperty,
+) : GraphicElement
