+++ remotecompose/lottie/src/main/java/com/google/android/horologist/remotecompose/lottie/renderer/Animation.kt
@@ -25,20 +25,20 @@ import androidx.compose.remote.creation.compose.state.lerp
 import androidx.compose.remote.creation.compose.state.rf
 import androidx.compose.remote.creation.compose.state.selectIfLt
 import com.google.android.horologist.remotecompose.lottie.LottieSettings
-import com.google.android.horologist.remotecompose.lottie.format.AnimatedBezierProperty
-import com.google.android.horologist.remotecompose.lottie.format.AnimatedPositionProperty
-import com.google.android.horologist.remotecompose.lottie.format.AnimatedScalarProperty
-import com.google.android.horologist.remotecompose.lottie.format.AnimatedVectorProperty
-import com.google.android.horologist.remotecompose.lottie.format.BaseBezierProperty
-import com.google.android.horologist.remotecompose.lottie.format.BasePositionProperty
-import com.google.android.horologist.remotecompose.lottie.format.BaseScalarProperty
-import com.google.android.horologist.remotecompose.lottie.format.BaseVectorProperty
-import com.google.android.horologist.remotecompose.lottie.format.BezierValue
-import com.google.android.horologist.remotecompose.lottie.format.ScalarKeyframeEasing
-import com.google.android.horologist.remotecompose.lottie.format.StaticBezierProperty
-import com.google.android.horologist.remotecompose.lottie.format.StaticPositionProperty
-import com.google.android.horologist.remotecompose.lottie.format.StaticScalarProperty
-import com.google.android.horologist.remotecompose.lottie.format.StaticVectorProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.AnimatedBezierProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.AnimatedPositionProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.AnimatedScalarProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.AnimatedVectorProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.BaseBezierProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.BasePositionProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.BaseScalarProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.BaseVectorProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.ScalarKeyframeEasing
+import com.google.android.horologist.remotecompose.lottie.format.properties.StaticBezierProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.StaticPositionProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.StaticScalarProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.StaticVectorProperty
+import com.google.android.horologist.remotecompose.lottie.format.values.BezierValue
 
 internal data class AnimationSegment(val startFrame: Float, val value: RemoteFloat)
