+++ remotecompose/lottie/src/test/java/com/google/android/horologist/remotecompose/lottie/AnimationTest.kt
@@ -18,9 +18,9 @@ package com.google.android.horologist.remotecompose.lottie
 
 import androidx.compose.remote.creation.compose.state.rf
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.google.android.horologist.remotecompose.lottie.format.AnimatedVectorProperty
-import com.google.android.horologist.remotecompose.lottie.format.StaticVectorProperty
-import com.google.android.horologist.remotecompose.lottie.format.VectorPropertyKeyframe
+import com.google.android.horologist.remotecompose.lottie.format.properties.AnimatedVectorProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.StaticVectorProperty
+import com.google.android.horologist.remotecompose.lottie.format.properties.VectorPropertyKeyframe
 import com.google.android.horologist.remotecompose.lottie.renderer.animateVector
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
